version 1.5
(no client changes)

SERVER

* allow for missing airport magnetic variations in APT.zip

* if exception decoding plate, just save in rejects file and keep going
  got exception decoding some line-drawn text military-only plate

* compensate for 2I0 airport miscoded with ICAO id K2I0 in APT.zip

* incorrectly decoded CI/CP airport status in APT.zip

* modernize difflatestaptdiags.sh to match DiffArptDgmCsvs.cs
  to detect changes in airport diagram georefs

* include some DecodePlate.java development scripts

* add review of deleted plates to iapreview.php
  to detect deleted iap plates vs rejections or missed downloads

+++ decosects/get_aptplates.sh
@@ -106,8 +106,17 @@ function downloadone
             if [ ! -f $pdftemp ]
             then
                 mkdir -p `dirname $pdftemp`
+                rm -f $pdftemp.$stream
                 wget -nv $pdfurl -O $pdftemp.$stream
-                mv -f $pdftemp.$stream $pdftemp
+                if [ ! -s $pdftemp.$stream ]
+                then
+                    rm -f $pdftemp.$stream
+                    wget -nv $pdfurl -O $pdftemp.$stream
+                fi
+                if [ -s $pdftemp.$stream ]
+                then
+                    mv -f $pdftemp.$stream $pdftemp
+                fi
             fi
 
             ## Maybe we already have the corresponding PNG file(s)
